Adding DogListActivity, ViewModel, Repository
