+++ app/src/main/java/com/example/dogdex/doglist/DogListViewModel.kt
@@ -1,0 +1,25 @@
+package com.example.dogdex.doglist
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
+import com.example.dogdex.Dog
+import kotlinx.coroutines.launch
+
+class DogListViewModel: ViewModel() {
+
+    private val _dogList = MutableLiveData<List<Dog>>()
+    val dogList: LiveData<List<Dog>>
+        get() = _dogList
+
+    private val dogRepository = DogRepository()
+    init {
+        downloadDogs()
+    }
+    private fun downloadDogs(){
+        viewModelScope.launch {
+            _dogList.value = dogRepository.downloadDogs()
+        }
+    }
+}
